requirements.txt
numpy==1.21.2
matplotlib==3.4.3
scikit-image==0.18.3
Pillow==8.3.2
opencv-python==4.5.3.56
PyYaml
jsonschema